Water Effects - Scene 3 test
Added in some water effects to test. These are in scene 3 for now.

To move - select the 'particles batch' object.
Note that the effects dont work outside of the range of 'effect camera', working on extending this to the same range as main camera.

There is also no gravity yet